refactor: Move buildConfiguration to ProjConf
