+++ services/src/test/scala/cromwell/pipeline/service/ProjectConfigurationServiceTest.scala
@@ -1,26 +1,38 @@
 package cromwell.pipeline.service
 
+import cats.data.NonEmptyList
 import cromwell.pipeline.datastorage.dao.repository.ProjectConfigurationRepository
 import cromwell.pipeline.datastorage.dao.utils.{ TestProjectUtils, TestUserUtils }
 import cromwell.pipeline.datastorage.dto._
 import cromwell.pipeline.model.wrapper.UserId
 import cromwell.pipeline.service.ProjectService.Exceptions.ProjectAccessDeniedException
+import cromwell.pipeline.womtool.WomTool
 import org.mockito.Mockito.when
 import org.scalatest.{ AsyncWordSpec, Matchers }
 import org.scalatestplus.mockito.MockitoSugar
 
-import java.nio.file.Paths
+import java.nio.file.{ Path, Paths }
 import scala.concurrent.Future
 
 class ProjectConfigurationServiceTest extends AsyncWordSpec with Matchers with MockitoSugar {
   private val projectService = mock[ProjectService]
   private val configurationRepository = mock[ProjectConfigurationRepository]
+  private val womTool = mock[WomTool]
+  private val projectVersioning = mock[ProjectVersioning[VersioningException]]
   private val dummyProject: Project = TestProjectUtils.getDummyProject()
-  private val configurationService = ProjectConfigurationService(configurationRepository, projectService)
+  private val configurationService =
+    ProjectConfigurationService(configurationRepository, projectService, womTool, projectVersioning)
   private val projectId: ProjectId = dummyProject.projectId
   private val userId: UserId = dummyProject.ownerId
   private val strangerId: UserId = TestUserUtils.getDummyUserId
   private val projectConfigurationId = ProjectConfigurationId.randomId
+  private val correctWdl = "task hello {}"
+  private val projectFilePath: Path = Paths.get("test.txt")
+  private val projectFileContent: ProjectFileContent = ProjectFileContent(correctWdl)
+  private val projectFile: ProjectFile = ProjectFile(projectFilePath, projectFileContent)
+  private val version: PipelineVersion = TestProjectUtils.getDummyPipeLineVersion()
+  private val optionVersion: Option[PipelineVersion] = Some(version)
+  private val errorMessage = "ERROR: miss bracket"
 
   private val projectFileConfiguration: ProjectFileConfiguration =
     ProjectFileConfiguration(Paths.get("/home/file"), List(FileParameter("nodeName", StringTyped(Some("hello")))))
@@ -157,6 +169,52 @@ class ProjectConfigurationServiceTest extends AsyncWordSpec with Matchers with M
           .deactivateLastByProjectId(projectId, strangerId)
           .failed
           .map(_ should have.message("Access denied. You  not owner of the project"))
+      }
+    }
+    "build configuration" should {
+
+      "return success message for request" taggedAs Service in {
+
+        when(configurationRepository.getAllByProjectId(projectId)).thenReturn(Future.successful(Seq()))
+        when(projectService.getUserProjectById(projectId, userId)).thenReturn(Future.successful(dummyProject))
+        when(projectVersioning.getFile(dummyProject, projectFile.path, optionVersion))
+          .thenReturn(Future.successful(Right(projectFile)))
+        when(womTool.inputsToList(projectFileContent.content)).thenReturn(Right(Nil))
+
+        configurationService
+          .buildConfiguration(projectId, projectFile.path, optionVersion, userId)
+          .map(
+            builtConfiguration =>
+              builtConfiguration shouldBe ProjectConfiguration(
+                id = builtConfiguration.id,
+                projectId = projectId,
+                active = true,
+                projectFileConfigurations = List(ProjectFileConfiguration(projectFile.path, Nil)),
+                version = ProjectConfigurationVersion.defaultVersion
+              )
+          )
+      }
+      "return error message for invalid file request" taggedAs Service in {
+        when(projectService.getUserProjectById(projectId, userId)).thenReturn(Future.successful(dummyProject))
+        when(projectVersioning.getFile(dummyProject, projectFile.path, optionVersion))
+          .thenReturn(Future.successful(Right(projectFile)))
+        when(womTool.inputsToList(projectFileContent.content)).thenReturn(Left(NonEmptyList(errorMessage, Nil)))
+
+        configurationService
+          .buildConfiguration(projectId, projectFile.path, optionVersion, userId)
+          .failed
+          .map(_ shouldBe ValidationError(List(errorMessage)))
+      }
+
+      "return error message for error request" taggedAs Service in {
+        when(projectService.getUserProjectById(projectId, userId)).thenReturn(Future.successful(dummyProject))
+        when(projectVersioning.getFile(dummyProject, projectFile.path, optionVersion))
+          .thenReturn(Future.successful(Left(VersioningException.FileException("404"))))
+
+        configurationService
+          .buildConfiguration(projectId, projectFile.path, optionVersion, userId)
+          .failed
+          .map(_ shouldBe VersioningException.FileException("404"))
       }
     }
   }
